Rename file for base class
