+++ src/figures/base.py
